+++ lib/rspec-cells.rb
@@ -10,27 +10,5 @@ module RSpec
         require 'rspec/cells'
       end
     end
-
-    module Caching
-      extend ActiveSupport::Concern
-
-      module ClassMethods
-        def enable_cell_caching!
-          before :each do
-            ActionController::Base.perform_caching = true
-          end
-          after :each do
-            ActionController::Base.perform_caching = false
-          end
-        end
-
-        def disable_cell_caching!
-          before :each do
-            ActionController::Base.perform_caching = false
-          end
-        end
-      end
-    end
-
   end
 end
