+++ lib/rspec/cells.rb
@@ -3,3 +3,4 @@ require 'rspec/core'
 require 'rspec/rails/adapters'
 require 'rspec/rails/example/rails_example_group'
 require 'rspec/cells/cell_example_group'
+require 'rspec/cells/caching'
