+++ spec/spec_helper.rb
@@ -8,7 +8,6 @@ require 'action_view/railtie'
 require 'rspec-cells'
 require 'rspec/rails'
 require 'cell/base'
-require 'cell/test_case'
 require 'rspec/cells'
 
 module RSpecCells
